+++ src/test/java/com/braintreegateway/unittest/DisputeGatewayTest.java
@@ -2,6 +2,7 @@ package com.braintreegateway.unittest;
 
 import com.braintreegateway.exceptions.NotFoundException;
 import com.braintreegateway.DisputeGateway;
+import com.braintreegateway.TextEvidenceRequest;
 
 import org.junit.Test;
 
@@ -36,7 +37,7 @@ public class DisputeGatewayTest {
             new DisputeGateway(null, null).addTextEvidence(null, "content");
             fail("DisputeGateway#addTextEvidence allowed null ID");
         } catch (NotFoundException exception) {
-            assertEquals(exception.getMessage(), "dispute with id \"null\" not found");
+            assertEquals(exception.getMessage(), "Dispute ID is required");
         }
     }
 
@@ -46,17 +47,29 @@ public class DisputeGatewayTest {
             new DisputeGateway(null, null).addTextEvidence(" ", "content");
             fail("DisputeGateway#addTextEvidence allowed empty string ID");
         } catch (NotFoundException exception) {
-            assertEquals(exception.getMessage(), "dispute with id \" \" not found");
+            assertEquals(exception.getMessage(), "Dispute ID is required");
+        }
+    }
+
+    @Test
+    public void addTextEvidenceNullTextEvidenceRequestRaisesIllegalArgumentException() {
+        try {
+            TextEvidenceRequest textEvidenceRequest = null;
+            new DisputeGateway(null, null).addTextEvidence("id", textEvidenceRequest);
+            fail("DisputeGateway#addTextEvidence allowed null content");
+        } catch (IllegalArgumentException exception) {
+            assertEquals(exception.getMessage(), "TextEvidenceRequest cannot be null");
         }
     }
 
     @Test
     public void addTextEvidenceNullContentRaisesIllegalArgumentException() {
         try {
-            new DisputeGateway(null, null).addTextEvidence("id", null);
+            String content = null;
+            new DisputeGateway(null, null).addTextEvidence("id", content);
             fail("DisputeGateway#addTextEvidence allowed null content");
         } catch (IllegalArgumentException exception) {
-            assertEquals(exception.getMessage(), "content cannot be empty");
+            assertEquals(exception.getMessage(), "Content cannot be empty");
         }
     }
 
@@ -66,7 +79,17 @@ public class DisputeGatewayTest {
             new DisputeGateway(null, null).addTextEvidence("id", " ");
             fail("DisputeGateway#addTextEvidence allowed empty string content");
         } catch (IllegalArgumentException exception) {
-            assertEquals(exception.getMessage(), "content cannot be empty");
+            assertEquals(exception.getMessage(), "Content cannot be empty");
+        }
+    }
+
+    @Test
+    public void addTextEvidenceRequestEmptyStringContentRaisesIllegalArgumentException() {
+        try {
+            new DisputeGateway(null, null).addTextEvidence("id", new TextEvidenceRequest().content(" "));
+            fail("DisputeGateway#addTextEvidence allowed empty string content");
+        } catch (IllegalArgumentException exception) {
+            assertEquals(exception.getMessage(), "Content cannot be empty");
         }
     }