+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add support for line_items
+* Add support for tagged evidence in DisputeGateway#addTextEvidence (Beta release)
 * Add support for raw Apple Pay processing
 * Add support for setting connect timeout (#56)